Refactor schnorr.go to schnorr_pubkey.go and schnorr_signature.go
